refactor(sessions): promote workspace/sessions to top-level feature
Phase 1 of 4 toward issue #181: move src/features/workspace/sessions/*
to src/features/sessions/* as a peer feature alongside workspace/.
Re-anchor cross-feature imports inside the moved files and update the
two WorkspaceView importers.

External deps (workspace/types, workspace/{utils,hooks}, StatusDot)
still point back into workspace/ via ../../workspace/... — resolved in
Phases 2-4.

Refs #181
