+++ src/features/sessions/components/Group.tsx
@@ -1,6 +1,6 @@
 import { type ReactElement, type ReactNode } from 'react'
 import { Reorder } from 'framer-motion'
-import type { Session } from '../../types'
+import type { Session } from '../../workspace/types'
 
 export interface GroupHeaderProps {
   label: string
