+++ src/features/workspace/WorkspaceView.tsx
@@ -1,11 +1,11 @@
 import type { ReactElement } from 'react'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { IconRail } from './components/IconRail'
-import { Tabs } from './sessions/components/Tabs'
+import { Tabs } from '../sessions/components/Tabs'
 import { Sidebar } from '../../components/sidebar/Sidebar'
 import { SidebarStatusHeader } from './components/SidebarStatusHeader'
 import { FileExplorer } from './components/panels/FileExplorer'
-import { List } from './sessions/components/List'
+import { List } from '../sessions/components/List'
 import { StatusBar } from './components/StatusBar'
 import { TerminalZone } from './components/TerminalZone'
 import BottomDrawer from './components/BottomDrawer'
